Refactor: 인벤토리 DTO 파일 이동 및 바뀐 파일명 반영
