+++ sever/src/main/java/com/example/rememberdokdo/Dto/Inventory/InventoryAddResponseDto.java
@@ -1,7 +1,7 @@
-package com.example.rememberdokdo.Dto;
+package com.example.rememberdokdo.Dto.Inventory;
 
 public class InventoryAddResponseDto {
     private Long sessionId; // 세션 식별자(넣을지 고민중)
     private String message; // 아이템 추가 성공 메시지
-    private InventoryItemDto inventoryItem; // 추가된 아이템 정보
+    private InventoryItemAddDto inventoryItem; // 추가된 아이템 정보
 }
